Fix setting row maker on nextset()
The test was broken and didn't test that the row_factory function was
really called.
